修改登录验证逻辑

+++ src/data/mutation/ArticleDelete.js
@@ -17,9 +17,20 @@ const articleDelete = {
 
     resolve: (root, args) => {
         return new Promise((resolve, reject) => {
-            BookModel.findByIdAndRemove(args.id, (err, book) => {
-                if (err) reject(err)
-                else resolve(book)
+            BookModel.findById(args.id, (err, book) => {
+                if (err) {
+                    reject(err)
+                } else if (book.author === root.request.user._doc._id) {
+                    BookModel.findByIdAndRemove(args.id, (err, book) => {
+                        if (err) {
+                            reject(err)
+                        } else {
+                            resolve(book)
+                        }
+                    });
+                } else {
+                    reject(new Error("您没有权限!"))
+                }
             });
         })
     }
